+++ src/components/Spinner/Spinner.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import spinner from './spinner.gif'
+import spinner from './assets/spinner.gif'
 import './Spinner.css'
 
 const Spinner = () => {
